Make glFieldButtonMenu multi-columned for large number of keys.
Make glFieldButtonMenu multi-columned for large number of keys.  I don't know
why, but I need to unrealize the old menu whenever reloading a new set of
keys -- otherwise, the new menu would not render properly.

Also, cleared the CAN_FOCUS and CAN_DEFAULT bits for the instance in
the object editor text edit page, so that it would not steal the focus
from the text editor.