Libglabels mods to support reorganization of template database
Added "equiv" attribute to "Template" tag.  The "equiv" attribute works
differently than the "Alias" tag: instead of defining a simple alias for
the current template struct, it says to use the equivalent part as the
prototype for this template.  The template could override (or append)
unique meta data for this part number.  The "Alias" tag is still supported,
but is deprecated.

Added a "product_url" attribute to the "Meta" tag.

Added a "vendor.xml" file which contains a database of vendors and their
primary url.  This file has only been populated with a few examples so far.

A new DTD file has been created to describe the above changes.

The "Avery 5160" family in "avery-us-templates.xml", has been organized in
the new scheme as an example.