Create a default frame if missing from template file
When parsing a template node, if there is no recognized frame node (e.g.
Label-rectangle), a default full-page frame with a 1x1 layout is created.
This prevents potential crashes if sharing a user '.glabels' directory
with future versions of glabels with new frame types, or in the case of
typos for hand generated templates.