Created new "New label dialog"
Created a new re-designed "New label dialog."  The dialog now shows meta-data
for the selected template, including hyperlinks to vendor and product
websites, if available.  The new dialog is more compact and thus should
be usable on small displays such as netbooks (probably not hand-helds).

The menu is somewhat clumbsy to navigate still -- will need to reorganize
to make navigation easier and more intuitive.