Do not explicitly set sidebar size.
Use 2 sizegroups to automatically fix sidebar size based on the size
of the widgets it contains -- even hidden ones.  The width was previously
set explicitly to prevent the sidebar from changing size whenever objects
are selected and unselected.  This has been problematic for compatability
with different themes, fonts, and translations.