Added new barcode drawing primitives.
Added a "Box" and a "String" barcode drawing primitives.  The box primitive
should be more natural for drawing actual boxes instead of lines.  The string
primitive allows an entire string to be rendered with a single interaction with
pango and cairo, rather than an interaction for each character.  There is no
need to try to implement kerning since pango will take care of the string as
a single layout.