Fixed problem with creation of 1st barcode, again.
Previous fix only temporarily masked the problem.  The problem is that
the first time the barcode tabs are made visible, the digits spin button
would emit a "changed" signal.  The fix is to connect to the "value-changed"
signal instead.

This commit also includes replacing the stop_signals flag with explicit
signal blockers.