Fix crash when encountering an invalid barcode backend
A file may have been saved from an instance of glabels that had a backend
enabled that is not enabled in the current instance of glabels.  This fix
accounts for this.  It is also possible that the glabels file was created
manually with an invalid backend ID or a typo.