Workaround for pango kerning bug.
Fixes bug #698777.  This is a workaround for pango-cairo bug #700592, which is
a regression of bug #341481 which was originally fixed back in 2009.  This
workaround works by rendering the font at device scale rather than world scale.
To make this work the actual font size is manually scaled appropriately to
compensate for this.