Auto-detect file encoding of CSV files and process accordingly.
Fixes sourceforge bug# 139. The changes scan any CSV file for a BOM
(Byte Order Mark) when it's opened. If one is found, then the getc
routine will use g_iconv to convert code sequences to utf-8 before
passing them up for processing. Since it checks for the UTF-8 BOM as
well as those for UTF-16 and UTF-32, this change also provides a way
in which a UTF-8 CSV file can be passed into GLabels on those systems
which aren't using UTF-8 as the system codeset. Currently, those
files are (it appears) assumed to be in the system code set, and so
are a bit problematic.